More clean up. Move modal type to separate file, getStyleForType, etc
