+++ src/components/AttachmentModal.js
@@ -4,10 +4,10 @@ import {
     View, TouchableOpacity, Text,
 } from 'react-native';
 import {withOnyx} from 'react-native-onyx';
-import Modal from './Modal';
+import MODAL_TYPE from './Modal/MODAL_TYPE';
+import ModalWithHeader from './ModalWithHeader';
 import AttachmentView from './AttachmentView';
 import styles, {colors} from '../styles/StyleSheet';
-import ModalHeader from './ModalHeader';
 import ONYXKEYS from '../ONYXKEYS';
 import addAuthTokenToURL from '../libs/addAuthTokenToURL';
 
@@ -66,45 +66,41 @@ class AttachmentModal extends Component {
 
         return (
             <>
-                <Modal
-                    type="centered"
+                <ModalWithHeader
+                    type={MODAL_TYPE.CENTERED}
                     onClose={() => this.setState({isModalOpen: false})}
                     isVisible={this.state.isModalOpen}
+                    title={this.props.title}
+                    backgroundColor={colors.componentBG}
                 >
-                    <View style={styles.modalViewContainer}>
-                        <ModalHeader
-                            title={this.props.title}
-                            onCloseButtonPress={() => this.setState({isModalOpen: false})}
-                        />
-                        <View style={styles.imageModalImageCenterContainer}>
-                            {this.state.sourceURL && (
-                                <AttachmentView sourceURL={sourceURL} file={this.state.file} />
-                            )}
-                        </View>
-
-                        {/* If we have an onConfirm method show a confirmation button */}
-                        {this.props.onConfirm && (
-                            <TouchableOpacity
-                                style={[styles.button, styles.buttonSuccess, styles.buttonConfirm]}
-                                underlayColor={colors.componentBG}
-                                onPress={() => {
-                                    this.props.onConfirm(this.state.file);
-                                    this.setState({isModalOpen: false});
-                                }}
-                            >
-                                <Text
-                                    style={[
-                                        styles.buttonText,
-                                        styles.buttonSuccessText,
-                                        styles.buttonConfirmText,
-                                    ]}
-                                >
-                                    Upload
-                                </Text>
-                            </TouchableOpacity>
+                    <View style={[styles.imageModalImageCenterContainer, {padding: 20}]}>
+                        {this.state.sourceURL && (
+                            <AttachmentView sourceURL={sourceURL} file={this.state.file} />
                         )}
                     </View>
-                </Modal>
+
+                    {/* If we have an onConfirm method show a confirmation button */}
+                    {this.props.onConfirm && (
+                        <TouchableOpacity
+                            style={[styles.button, styles.buttonSuccess, styles.buttonConfirm]}
+                            underlayColor={colors.componentBG}
+                            onPress={() => {
+                                this.props.onConfirm(this.state.file);
+                                this.setState({isModalOpen: false});
+                            }}
+                        >
+                            <Text
+                                style={[
+                                    styles.buttonText,
+                                    styles.buttonSuccessText,
+                                    styles.buttonConfirmText,
+                                ]}
+                            >
+                                Upload
+                            </Text>
+                        </TouchableOpacity>
+                    )}
+                </ModalWithHeader>
                 {this.props.children({
                     displayFileInModal: ({file}) => {
                         if (file instanceof File) {
