+++ src/components/Modal/MODAL_TYPE.js
@@ -1,0 +1,3 @@
+export default {
+    CENTERED: 'centered',
+};
