+++ src/components/Modal/getStyleForType.js
@@ -1,0 +1,66 @@
+import MODAL_TYPE from './MODAL_TYPE';
+import {colors} from '../../styles/StyleSheet';
+import variables from '../../styles/variables';
+
+export default (type, windowDimensions) => {
+    const isSmallScreen = windowDimensions.width < variables.mobileResponsiveWidthBreakpoint;
+
+    let modalStyle;
+    let modalContainerStyle;
+    let swipeDirection;
+    let animationIn;
+    let animationOut;
+    let needsSafeAreaPadding = false;
+
+    switch (type) {
+        case MODAL_TYPE.CENTERED:
+            // A centered modal is one that has a visible backdrop
+            // and can be dismissed by clicking outside of the modal.
+            // This modal should take up the entire visible area when
+            // viewed on a smaller device (e.g. mobile or mobile web).
+            modalStyle = {
+                margin: 0,
+                alignItems: 'center',
+            };
+            modalContainerStyle = {
+                // Shadow Styles
+                shadowColor: colors.black,
+                shadowOffset: {
+                    width: 0,
+                    height: 0,
+                },
+                shadowOpacity: 0.3,
+                shadowRadius: 20,
+
+                flex: 1,
+                marginTop: isSmallScreen ? 0 : 20,
+                marginBottom: isSmallScreen ? 0 : 20,
+                borderRadius: isSmallScreen ? 0 : 12,
+                borderWidth: isSmallScreen ? 1 : 0,
+                borderColor: colors.border,
+                overflow: 'hidden',
+                width: isSmallScreen ? '100%' : windowDimensions.width - 40,
+            };
+
+            swipeDirection = undefined;
+            animationIn = 'fadeIn';
+            animationOut = 'fadeOut';
+            needsSafeAreaPadding = true;
+            break;
+        default:
+            modalStyle = {};
+            modalContainerStyle = {};
+            swipeDirection = 'down';
+            animationIn = 'slideInUp';
+            animationOut = 'slideOutDown';
+    }
+
+    return {
+        modalStyle,
+        modalContainerStyle,
+        swipeDirection,
+        animationIn,
+        animationOut,
+        needsSafeAreaPadding,
+    };
+};
