+++ src/styles/variables.js
@@ -1,0 +1,5 @@
+export default {
+    modalHeaderBarHeight: 73,
+    safeInsertPercentage: 0.7,
+    mobileResponsiveWidthBreakpoint: 1000,
+};
